requirements.txt
chromedriver_autoinstaller==0.6.4
python-dotenv==1.0.1
selenium==4.26.0
